# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
/**
 * Typed error codes for all application errors
 */
export type AppErrorCode = (typeof AppErrorCode)[keyof typeof AppErrorCode];

export const AppErrorCode = {
  BAD_REQUEST: "BAD_REQUEST",
  UNAUTHORIZED: "UNAUTHORIZED",
  FORBIDDEN: "FORBIDDEN",
  NOT_FOUND: "NOT_FOUND",
  CONFLICT: "CONFLICT",
  UNPROCESSABLE: "UNPROCESSABLE",
  INTERNAL_ERROR: "INTERNAL_ERROR",
  VALIDATION_ERROR: "VALIDATION_ERROR",
  ENTITY_TYPE_IN_USE: "ENTITY_TYPE_IN_USE",
} as const;

/**
 * Optional additional error context (field errors, etc.)
 * @nullable
 */
export type AppErrorDetails = { [key: string]: unknown } | null;

/**
 * Structured error object included in error responses
 */
export interface AppError {
  code: AppErrorCode;
  /** Human-readable error message */
  message: string;
  /**
   * Optional additional error context (field errors, etc.)
   * @nullable
   */
  details?: AppErrorDetails;
}

/**
 * Optional metadata attached to responses (pagination, timing, etc.)
 */
export interface ApiResponseMeta {
  [key: string]: unknown;
}

/**
 * A catalog entity type definition
 */
export interface EntityType {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  isSystemSeed: boolean;
  fieldCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateEntityTypeInput {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /**
   * @maxLength 500
   * @nullable
   */
  description?: string | null;
}

export interface UpdateEntityTypeInput {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name?: string;
  /**
   * @maxLength 500
   * @nullable
   */
  description?: string | null;
}

export interface DeleteEntityTypeResult {
  deleted: boolean;
}

export interface EntityTypeApiResponse {
  data: EntityType | null;
  error: AppError | null;
  meta?: ApiResponseMeta | null;
}

export interface EntityTypeListApiResponse {
  data: EntityType[] | null;
  error: AppError | null;
  meta?: ApiResponseMeta | null;
}

export interface DeleteEntityTypeApiResponse {
  data: DeleteEntityTypeResult | null;
  error: AppError | null;
  meta?: ApiResponseMeta | null;
}

export interface HealthStatus {
  status: string;
}

/**
 * Envelope for health check responses
 */
export interface HealthCheckApiResponse {
  data: HealthStatus | null;
  error: AppError | null;
  meta?: ApiResponseMeta | null;
}

/**
 * Generic error envelope returned on failure
 */
export interface ErrorApiResponse {
  /** @nullable */
  data: null;
  error: AppError;
  meta?: ApiResponseMeta | null;
}