# lib/api-zod/src/generated/types/appErrorCode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Typed error codes for all application errors
 */
export type AppErrorCode = (typeof AppErrorCode)[keyof typeof AppErrorCode];

export const AppErrorCode = {
  BAD_REQUEST: "BAD_REQUEST",
  UNAUTHORIZED: "UNAUTHORIZED",
  FORBIDDEN: "FORBIDDEN",
  NOT_FOUND: "NOT_FOUND",
  CONFLICT: "CONFLICT",
  UNPROCESSABLE: "UNPROCESSABLE",
  INTERNAL_ERROR: "INTERNAL_ERROR",
  VALIDATION_ERROR: "VALIDATION_ERROR",
  ENTITY_TYPE_IN_USE: "ENTITY_TYPE_IN_USE",
} as const;