# lib/api-zod/src/generated/types/deleteEntityTypeApiResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApiResponseMeta } from "./apiResponseMeta";
import type { AppError } from "./appError";
import type { DeleteEntityTypeResult } from "./deleteEntityTypeResult";

export interface DeleteEntityTypeApiResponse {
  data: DeleteEntityTypeResult | null;
  error: AppError | null;
  meta?: ApiResponseMeta | null;
}