# lib/api-zod/src/generated/types/entityType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * A catalog entity type definition
 */
export interface EntityType {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  isSystemSeed: boolean;
  fieldCount: number;
  createdAt: Date;
  updatedAt: Date;
}