# lib/api-zod/src/generated/types/entityTypeApiResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApiResponseMeta } from "./apiResponseMeta";
import type { AppError } from "./appError";
import type { EntityType } from "./entityType";

export interface EntityTypeApiResponse {
  data: EntityType | null;
  error: AppError | null;
  meta?: ApiResponseMeta | null;
}