# lib/api-zod/src/generated/types/updateEntityTypeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateEntityTypeInput {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name?: string;
  /**
   * @maxLength 500
   * @nullable
   */
  description?: string | null;
}